Move search result data to seprate file
to "declutter" the SearchResultPage component
